Change some Actions to getters
