+++ src/ADempiere/modules/persistence/store/Data/actions.ts
@@ -26,7 +26,6 @@ import {
   IPrivateAccessDataExtended,
   IRecordObjectListFromCriteria,
   IRecordSelectionData,
-  ISelectionToServerData,
   KeyValueData
 } from '../../PersistenceType'
 import language from '@/ADempiere/shared/lang'
@@ -1108,90 +1107,6 @@ export const actions: BusinessDataActionTree = {
                     `Error unlock private access: ${error.message}. Code: ${error.code}.`
         )
       })
-  },
-  /**
-     * Getter converter selection data record in format
-     * @param {string} containerUuid
-     * @param {array}  selection
-     * [{
-     *    selectionId: keyColumn Value,
-     *    selectionValues: [{ columnName, value }]
-     * }]
-     */
-  getSelectionToServer: (context: BusinessDataActionContext) => (parameters: {
-        containerUuid: string
-        selection?: IKeyValueObject[]
-    }): ISelectionToServerData[] => {
-    let { selection = parameters.selection || [] } = parameters
-    const { containerUuid } = parameters
-    const selectionToServer: ISelectionToServerData[] = []
-
-    const withOut: string[] = ['isEdit', 'isSendToServer']
-
-    if (selection.length <= 0) {
-      selection = context.getters.getDataRecordSelection(containerUuid)
-    }
-    if (selection.length) {
-      const { fieldsList, keyColumn } = <IPanelDataExtended>(
-                context.rootGetters[Namespaces.Panel + '/' + 'getPanel'](containerUuid)
-            )
-            // reduce list
-      const fieldsListSelection: IFieldDataExtendedUtils[] = fieldsList.filter(
-        (itemField: IFieldDataExtendedUtils) => {
-          return itemField.isIdentifier || itemField.isUpdateable
-        }
-      )
-
-      selection.forEach((itemRow: IKeyValueObject) => {
-        const records: KeyValueData[] = []
-
-        Object.keys(itemRow).forEach((key: string) => {
-          if (
-            !key.includes('DisplayColumn') &&
-                        !withOut.includes(key)
-          ) {
-            // evaluate metadata attributes before to convert
-            const field:
-                            | IFieldDataExtendedUtils
-                            | undefined = fieldsListSelection.find(
-                              (itemField: IFieldDataExtendedUtils) =>
-                                itemField.columnName === key
-                            )
-            if (field) {
-              records.push({
-                // columnName: key,
-                key,
-                value: itemRow[key]
-              })
-            }
-          }
-        })
-
-        selectionToServer.push({
-          selectionId: itemRow[keyColumn!],
-          selectionValues: records
-        })
-      })
-    }
-    return selectionToServer
-  },
-  getRowData: (context: BusinessDataActionContext) => (data: {
-        containerUuid: string
-        recordUuid?: string
-        index: number
-    }) => {
-    const { recordUuid, containerUuid, index } = data
-    const recordsList: any[] = context.getters.getDataRecordsList(
-      containerUuid
-    )
-    if (index) {
-      return recordsList[index]
-    }
-    return recordsList.find((itemData: any) => {
-      if (itemData.UUID === recordUuid) {
-        return true
-      }
-    })
   },
   resetStateBusinessData(context: BusinessDataActionContext) {
     const { commit } = context
